+++ src/DotPulsar/ProducerOptions.cs
@@ -66,6 +66,7 @@ public sealed class ProducerOptions<TMessage>
         EncryptionKeys = DefaultEncryptionKeys;
         CryptoKeyReader = DefaultCryptoKeyReader;
         CryptoFailureAction = DefaultCryptoFailureAction;
+        ProducerProperties = [];
     }
 
     /// <summary>
@@ -132,4 +133,9 @@ public sealed class ProducerOptions<TMessage>
     /// Set the max size of the queue holding the messages pending to receive an acknowledgment from the broker.
     /// </summary>
     public uint MaxPendingMessages { get; set; }
+
+    /// <summary>
+    /// Add/Set the producers's properties. This is optional.
+    /// </summary>
+    public Dictionary<string, string> ProducerProperties { get; set; }
 }